params: Add GetType{Path,Query,Header,Cookie} param functions (#45)
* params: Add Get*From param functions, support Cookies

Currently params.Get* will look for a parameter in the request path
values, query string and Headers in that order. There is no way to be
specific in what kind of parameter you want to fetch.

This commit adds two new features:

* params.Get*From functions, which take an extra source parameter that
  allows you to specific which location you want to get a parameter
  from, with no fallback.

* Cookie support (in the Get*From functions only).

This is motivated by a desire to support Cookies in gofoji's openapi
request parsing, but to do so without accidentally leaking cookies into
any existing users of the param package.

* Use function per parameter source

* More comprehensive tests

In order to test comprehensively, we need to run 4 or so tests for each
param source, for each type. That's 176 very repetative tests. So this
is implemented with a rather difficult to read matrix test method.